requirements.txt
asgiref==3.4.1
cachelib==0.4.1
click==8.0.3
dataclasses==0.6
debugpy==1.5.1
dj-database-url==0.5.0
Django==3.2
django-debug-toolbar==3.2.2
Dumper==1.2.0
Faker==9.9.0
Flask==2.0.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pymongo==4.0.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.28
SQLAlchemy-Utils==0.37.9
sqlparse==0.4.2
text-unidecode==1.3
typing_extensions==4.0.0
Werkzeug==2.0.2
whitenoise==5.3.0
zipp==3.6.0
